PaddleOCR/requirements.txt
paddlepaddle
paddleocr
flask
werkzeug
pillow